Rework test name listener to correct name in allure report
