+++ aquality-appium-mobile-template-cucumber/src/test/java/aquality/appium/mobile/template/cucumber/features/Demo.feature
@@ -1,6 +1,4 @@
 Feature: Demo
-  Background:
-    Given Allure test name has been updated by the platform parameter
 
   @demo
   Scenario: I try to login with invalid credentials
