+++ aquality-appium-mobile-template/src/main/resources/META-INF/services/io.qameta.allure.listener.TestLifecycleListener
@@ -1,0 +1,1 @@
+aquality.appium.mobile.template.utilities.AllureListener
